fix(SOCRadar): rename FeedSOCRadarThreatFeed v2 files to FeedSOCRadarThreatFeedV2
The v2 integration YAML already declared id/name FeedSOCRadarThreatFeedV2, but its
directory and file basenames were still FeedSOCRadarThreatFeed. That collided with the
legacy integration in Packs/FeedSOCRadarThreatFeed, which shares the same module
basename and the same docker image, so demisto-sdk batched both into one mypy run and
mypy aborted with "Duplicate module named FeedSOCRadarThreatFeed" (exit code 2).

Renamed the v2 directory and files to match the integration name, updated the
.pack-ignore file reference and the test module imports, and bumped the pack to 2.3.3.
