+++ Packs/SOCRadar/Integrations/FeedSOCRadarThreatFeedV2/test_data/check_auth_response_auth_error.json
