+++ Packs/SOCRadar/ReleaseNotes/2_3_3.md
@@ -1,0 +1,6 @@
+
+#### Integrations
+
+##### SOCRadar Threat Feed v2
+
+- Renamed the integration files to match the integration name.
