+++ Packs/SOCRadar/pack_metadata.json
@@ -3,7 +3,7 @@
     "created": "2021-08-05T16:43:31Z",
     "description": "Streamline remediation of alerts and incidents with enhanced multi-tenant capabilities. Easily manage and automate security operations across multiple tenants, including for enrichment threat intelligence, reputation checking, and IoC feeds",
     "support": "partner",
-    "currentVersion": "2.3.2",
+    "currentVersion": "2.3.3",
     "author": "SOCRadar Cyber Intelligence Inc.",
     "defaultDataSource": "SOCRadarIncidentsV4",
     "url": "www.socradar.io",
